+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Adds [`Objectware.forEach`][forEach].
 - Adds [`Objectware.eachOwn`][eachOwn].
 - Adds [`Objectware.forEachOwn`][forEachOwn].
+- Adds [`Objectware.isInOwn`][isInOwn].
 
 [isOwnEmpty]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.isOwnEmpty
 [has]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.has
@@ -14,6 +15,7 @@
 [forEach]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.forEach
 [eachOwn]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.eachOwn
 [forEachOwn]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.forEachOwn
+[isInOwn]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.isInOwn
 
 ## 1.4.0 (Feb 8, 2015)
 - Adds [`Objectware.reject`][reject].